Add rm command.
rm command moves files to trash directory. It creates a new deletion
'instance' in trash dir. This avoids the problem of deleting two files
with the same name.
It also enables easy restoration of the file to its original location.
This is still not implemented, it will come in a later commit.

Command checks if there is a trash directory. If it doesn't it will try
to create it.

Note: some functions are defined in this file at the moment, but they
will be moved out as soon as some other command/function needs them.